refactor: Controller 예약, 시간 분리
