+++ src/main/java/roomescape/controller/ReservationController.java
@@ -9,9 +9,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.Reservation;
-import roomescape.domain.ReservationTime;
 import roomescape.dto.CreateReservationRequest;
-import roomescape.dto.CreateReservationTimeRequest;
 import roomescape.service.ReservationService;
 
 @RestController
@@ -46,21 +44,4 @@ public class ReservationController {
         return ResponseEntity.ok().build();
     }
 
-    @GetMapping("/times")
-    public ResponseEntity<List<ReservationTime>> readTimes() {
-        return ResponseEntity.ok().body(reservationService.getReservationTimes());
-    }
-
-    @PostMapping("/times")
-    public ResponseEntity<ReservationTime> createTime(
-            @RequestBody CreateReservationTimeRequest createReservationTimeRequest) {
-        ReservationTime createdTime = reservationService.createReservationTime(createReservationTimeRequest.startAt());
-        return ResponseEntity.ok().body(createdTime);
-    }
-
-    @DeleteMapping("/times/{id}")
-    public ResponseEntity<Void> deleteTime(@PathVariable Long id) {
-        reservationService.deleteReservationTime(id);
-        return ResponseEntity.ok().build();
-    }
 }
