+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -1,0 +1,41 @@
+package roomescape.controller;
+
+import java.util.List;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.domain.ReservationTime;
+import roomescape.dto.CreateReservationTimeRequest;
+import roomescape.service.ReservationService;
+
+@RestController
+public class ReservationTimeController {
+
+    private final ReservationService reservationService;
+
+    public ReservationTimeController(ReservationService reservationService) {
+        this.reservationService = reservationService;
+    }
+
+    @GetMapping("/times")
+    public ResponseEntity<List<ReservationTime>> readTimes() {
+        return ResponseEntity.ok().body(reservationService.getReservationTimes());
+    }
+
+    @PostMapping("/times")
+    public ResponseEntity<ReservationTime> createTime(
+            @RequestBody CreateReservationTimeRequest createReservationTimeRequest) {
+        ReservationTime createdTime = reservationService.createReservationTime(createReservationTimeRequest.startAt());
+        return ResponseEntity.ok().body(createdTime);
+    }
+
+    @DeleteMapping("/times/{id}")
+    public ResponseEntity<Void> deleteTime(@PathVariable Long id) {
+        reservationService.deleteReservationTime(id);
+        return ResponseEntity.ok().build();
+    }
+}
